requirements.txt
numpy-2.0.2
pandas-2.2.3
pytz-2024.2
tzdata-2024.2
beautifulsoup4-4.12.3
bs4-0.0.2
soupsieve-2.6
certifi-2024.12.14
charset-normalizer-3.4.1
idna-3.10
requests-2.32.3
urllib3-2.3.0
contourpy-1.3.0
cycler-0.12.1
fonttools-4.56.0
importlib-resources-6.5.2
kiwisolver-1.4.7
matplotlib-3.9.4
pillow-11.1.0
pyparsing-3.2.1
seaborn-0.13.2